[EMBR-6046] Split functionality from core/index.ts into separate files
