+++ packages/core/src/index.ts
@@ -1,80 +1,19 @@
 "use strict";
+
 import {Platform} from "react-native";
 
 import * as embracePackage from "../package.json";
 
-import {generateStackTrace, handleGlobalError} from "./utils/ErrorUtil";
-import {logIfComponentError} from "./utils/ComponentError";
-import {ApplyInterceptorStrategy} from "./networkInterceptors/ApplyInterceptor";
-import {
-  LogSeverity,
-  SessionStatus,
-  SDKConfig,
-  MethodType,
-  LogProperties,
-} from "./interfaces/common";
-import {
-  getNetworkSDKInterceptorProvider,
-  NETWORK_INTERCEPTOR_TYPES,
-} from "./interfaces/NetworkMonitoring";
+import {setJavaScriptPatch} from "./utils/javascript";
+import {buildPackageVersion, isObjectNonEmpty} from "./utils/global";
+import {handleError, handleGlobalError} from "./utils/error";
+import {SDKConfig} from "./interfaces";
 import {EmbraceManagerModule} from "./EmbraceManagerModule";
 
 const reactNativeVersion = require("react-native/Libraries/Core/ReactNativeVersion.js");
 const tracking = require("promise/setimmediate/rejection-tracking");
 
-const STACK_LIMIT = 200;
 const UNHANDLED_PROMISE_REJECTION_PREFIX = "Unhandled promise rejection";
-
-const noOp = () => {};
-
-// will cover unhandled errors, js crashes
-const handleError = async (error: Error, callback: () => void) => {
-  if (!(error instanceof Error)) {
-    console.warn("[Embrace] error must be of type Error");
-    return;
-  }
-
-  const {name, message, stack = ""} = error;
-
-  logIfComponentError(error);
-
-  // same as error.name? why is it pulled differently?
-  const errorType = error.constructor.name;
-
-  // truncating stacktrace to 200 lines
-  const stTruncated = stack.split("\n").slice(0, STACK_LIMIT);
-
-  // specifically for iOS for now, the same formatting is done in the Android layer
-  // in the future Android will get rid of all related to js and use this format as well
-  const iosStackTrace = JSON.stringify({
-    n: name,
-    m: message,
-    t: errorType,
-    // removing the Type from the first part of the stacktrace.
-    st: stTruncated.slice(1, stTruncated.length).join("\n"),
-  });
-
-  let logException;
-  try {
-    logException = await EmbraceManagerModule.logUnhandledJSException(
-      name,
-      message,
-      errorType,
-      Platform.OS === "android" ? stTruncated.join("\n") : iosStackTrace,
-    );
-  } catch {
-    logException = false;
-  }
-
-  if (!logException) {
-    console.warn("[Embrace] Failed to log exception");
-  }
-
-  callback();
-};
-
-const isObjectNonEmpty = (obj?: object): boolean =>
-  Object.keys(obj || {}).length > 0;
 
 const initialize = async ({
   patch,
@@ -131,7 +70,7 @@ const initialize = async ({
     isObjectNonEmpty(reactNativeVersion.version)
   ) {
     EmbraceManagerModule.setReactNativeVersion(
-      buildVersionStr(reactNativeVersion.version),
+      buildPackageVersion(reactNativeVersion.version),
     );
   }
 
@@ -191,259 +130,20 @@ const initialize = async ({
         stackTrace,
       );
     },
-    onHandled: noOp,
+    onHandled: () => {},
   });
 
   // If there are no errors, it will return true
   return Promise.resolve(true);
 };
 
-const buildVersionStr = ({
-  major,
-  minor,
-  patch,
-  prerelease,
-}: {
-  major: string;
-  minor: string;
-  patch: string;
-  prerelease: string | null;
-}): string => {
-  const versionStr = `${major || "0"}.${minor || "0"}.${patch || "0"}`;
-  return prerelease ? `${versionStr}.${prerelease}` : versionStr;
-};
+export * from "./utils/log";
+export * from "./utils/session";
+export * from "./utils/breadcrumb";
+export * from "./utils/views";
+export * from "./utils/network";
+export * from "./utils/user";
+export * from "./utils/error";
+export * from "./utils/component";
 
-export const setUserIdentifier = (userIdentifier: string): Promise<boolean> => {
-  return EmbraceManagerModule.setUserIdentifier(userIdentifier);
-};
-
-export const clearUserIdentifier = (): Promise<boolean> => {
-  return EmbraceManagerModule.clearUserIdentifier();
-};
-
-export const setUsername = (username: string): Promise<boolean> => {
-  return EmbraceManagerModule.setUsername(username);
-};
-
-export const clearUsername = (): Promise<boolean> => {
-  return EmbraceManagerModule.clearUsername();
-};
-
-export const setUserEmail = (userEmail: string): Promise<boolean> => {
-  return EmbraceManagerModule.setUserEmail(userEmail);
-};
-
-export const clearUserEmail = (): Promise<boolean> => {
-  return EmbraceManagerModule.clearUserEmail();
-};
-
-export const addBreadcrumb = (message: string): Promise<boolean> => {
-  return EmbraceManagerModule.addBreadcrumb(message);
-};
-
-export const logScreen = (screenName: string): Promise<boolean> => {
-  return EmbraceManagerModule.addBreadcrumb(`Opening screen [${screenName}]`);
-};
-
-export const addUserPersona = (persona: string): Promise<boolean> => {
-  return EmbraceManagerModule.addUserPersona(persona);
-};
-
-export const clearUserPersona = (persona: string): Promise<boolean> => {
-  return EmbraceManagerModule.clearUserPersona(persona);
-};
-
-export const clearAllUserPersonas = (): Promise<boolean> => {
-  return EmbraceManagerModule.clearAllUserPersonas();
-};
-
-const logMessage = (
-  message: string,
-  severity: LogSeverity = "error",
-  // Android Native method is handling the null case
-  // iOS Native method is waiting for a non-nullable object
-  properties: LogProperties = {},
-  includeStacktrace = true,
-): Promise<boolean> => {
-  const stackTrace =
-    // `"info"` are not supposed to send stack traces
-    // this is also restricted in the Native layers
-    includeStacktrace && severity !== "info" ? generateStackTrace() : "";
-
-  if (properties === null) {
-    console.warn(
-      "[Embrace] `properties` is null. It should be an object of type `Properties`. Native layer will ignore this value.",
-    );
-  }
-
-  return EmbraceManagerModule.logMessageWithSeverityAndProperties(
-    message,
-    severity,
-    properties,
-    stackTrace,
-    includeStacktrace,
-  );
-};
-
-const logInfo = (message: string): Promise<boolean> => {
-  // `"info"` logs are not supposed to send stack traces as per Product decision
-  // this is also restricted in the Native layers
-  return logMessage(message, "info", undefined, false);
-};
-
-const logWarning = (
-  message: string,
-  includeStacktrace = true,
-): Promise<boolean> => {
-  return logMessage(message, "warning", undefined, includeStacktrace);
-};
-
-const logError = (
-  message: string,
-  includeStacktrace = true,
-): Promise<boolean> => {
-  return logMessage(message, "error", undefined, includeStacktrace);
-};
-
-const logHandledError = (
-  error: Error,
-  properties: LogProperties = {},
-): Promise<boolean> => {
-  if (error instanceof Error) {
-    const {stack, message} = error;
-    return EmbraceManagerModule.logHandledError(message, stack, properties);
-  }
-
-  return Promise.resolve(false);
-};
-
-export const startView = (view: string): Promise<string | boolean> => {
-  return EmbraceManagerModule.startView(view);
-};
-
-export const endView = (spanId: string): Promise<boolean> => {
-  return EmbraceManagerModule.endView(spanId);
-};
-
-export const setJavaScriptPatch = (patch: string) => {
-  return EmbraceManagerModule.setJavaScriptPatchNumber(patch);
-};
-
-export const setJavaScriptBundlePath = (path: string) => {
-  return EmbraceManagerModule.setJavaScriptBundlePath(path);
-};
-
-export const addSessionProperty = (
-  key: string,
-  value: string,
-  permanent: boolean,
-): Promise<boolean> => {
-  return EmbraceManagerModule.addSessionProperty(key, value, permanent);
-};
-
-export const removeSessionProperty = (key: string) => {
-  return EmbraceManagerModule.removeSessionProperty(key);
-};
-
-export const endSession = () => {
-  return EmbraceManagerModule.endSession();
-};
-
-export const setUserAsPayer = (): Promise<boolean> => {
-  return EmbraceManagerModule.setUserAsPayer();
-};
-
-export const clearUserAsPayer = (): Promise<boolean> => {
-  return EmbraceManagerModule.clearUserAsPayer();
-};
-
-export const recordNetworkRequest = (
-  url: string,
-  httpMethod: MethodType,
-  startInMillis: number,
-  endInMillis: number,
-  bytesSent?: number,
-  bytesReceived?: number,
-  statusCode?: number,
-): Promise<boolean> => {
-  return EmbraceManagerModule.logNetworkRequest(
-    url,
-    httpMethod,
-    startInMillis,
-    endInMillis,
-    bytesSent || -1,
-    bytesReceived || -1,
-    statusCode || -1,
-  );
-};
-
-export const logNetworkClientError = (
-  url: string,
-  httpMethod: MethodType,
-  startInMillis: number,
-  endInMillis: number,
-  errorType: string,
-  errorMessage: string,
-): Promise<boolean> => {
-  return EmbraceManagerModule.logNetworkClientError(
-    url,
-    httpMethod,
-    startInMillis,
-    endInMillis,
-    errorType,
-    errorMessage,
-  );
-};
-
-export const getLastRunEndState = (): Promise<SessionStatus> =>
-  EmbraceManagerModule.getLastRunEndState();
-
-export const getDeviceId = (): Promise<string> =>
-  EmbraceManagerModule.getDeviceId();
-
-export const getCurrentSessionId = (): Promise<string> =>
-  EmbraceManagerModule.getCurrentSessionId();
-
-export const applyNetworkInterceptors = (
-  networkSDKInstance: NETWORK_INTERCEPTOR_TYPES,
-): Promise<boolean> => {
-  if (!networkSDKInstance) {
-    console.warn(
-      `[Embrace] The Axios instance was not provided. Interceptor was not applied.`,
-    );
-    return createFalsePromise();
-  }
-
-  const networkProviderSDK =
-    getNetworkSDKInterceptorProvider(networkSDKInstance);
-
-  if (!networkProviderSDK) {
-    console.warn(
-      `[Embrace] The provider is not supported. Interceptor was not applied.`,
-    );
-    return createFalsePromise();
-  }
-
-  const {applyInterceptor} = ApplyInterceptorStrategy[networkProviderSDK];
-
-  applyInterceptor(networkSDKInstance);
-  return createTruePromise();
-};
-
-const createFalsePromise = (): Promise<boolean> => {
-  return new Promise(resolve => {
-    setTimeout(() => {
-      resolve(false);
-    }, 0);
-  });
-};
-
-const createTruePromise = (): Promise<boolean> => {
-  return new Promise(resolve => {
-    setTimeout(() => {
-      resolve(true);
-    }, 0);
-  });
-};
-
-export {initialize, logError, logHandledError, logInfo, logWarning, logMessage};
+export {initialize};
