--- packages/core/src/interfaces/Config.ts
@@ -1,7 +1,0 @@
-/**
- * Defines the SDK configuration that can be passed into `initialize`
- */
-
-import {SDKConfig, IOSConfig} from "./common";
-
-export {SDKConfig, IOSConfig};
