+++ packages/core/src/utils/component.ts
