+++ packages/core/src/utils/network.ts
@@ -1,0 +1,42 @@
+import {MethodType} from "../interfaces";
+import {EmbraceManagerModule} from "../EmbraceManagerModule";
+
+const recordNetworkRequest = (
+  url: string,
+  httpMethod: MethodType,
+  startInMillis: number,
+  endInMillis: number,
+  bytesSent?: number,
+  bytesReceived?: number,
+  statusCode?: number,
+): Promise<boolean> => {
+  return EmbraceManagerModule.logNetworkRequest(
+    url,
+    httpMethod,
+    startInMillis,
+    endInMillis,
+    bytesSent || -1,
+    bytesReceived || -1,
+    statusCode || -1,
+  );
+};
+
+const logNetworkClientError = (
+  url: string,
+  httpMethod: MethodType,
+  startInMillis: number,
+  endInMillis: number,
+  errorType: string,
+  errorMessage: string,
+): Promise<boolean> => {
+  return EmbraceManagerModule.logNetworkClientError(
+    url,
+    httpMethod,
+    startInMillis,
+    endInMillis,
+    errorType,
+    errorMessage,
+  );
+};
+
+export {recordNetworkRequest, logNetworkClientError};
