+++ packages/react-native-otlp/package.json
@@ -18,7 +18,7 @@
     "test": "__tests__"
   },
   "scripts": {
-    "js:shared": "mkdir -p ./src/interfaces/common && cp -r ../../common/interfaces/* ./src/interfaces/common",
+    "js:shared": "mkdir -p ./src/interfaces/common && cp -r ../../common/js/* ./src/interfaces/common",
     "ios:shared": "mkdir -p ./ios/common && cp -r ../../common/ios/* ./ios/common/",
     "build": "yarn run js:shared && yarn run ios:shared && tsc",
     "android:test": "cd test-project && yarn && cd android && ./gradlew :app:detek && ./gradlew :app:testDebugUnitTest && cd ../../",
