+++ packages/react-native-otlp/src/interfaces/index.ts
@@ -1,5 +1,3 @@
-import {IOSConfig, AndroidConfig} from "./common/index";
-
 interface ExporterConfig {
   endpoint: string;
   headers?: {key: string; token: string}[];
@@ -11,4 +9,6 @@ interface OTLPExporterConfig {
   traceExporter?: ExporterConfig;
 }
 
-export {IOSConfig, AndroidConfig, OTLPExporterConfig, ExporterConfig};
+export * from "./common/types";
+export * from "./common/interfaces";
+export {OTLPExporterConfig, ExporterConfig};
